refactor cart logic

+++ src/modules/Cart/CartContent.tsx
@@ -1,9 +1,8 @@
+import { useAppDispatch, useAppSelector } from '@/redux/hooks';
+import { deleteItem, getFilteredCart } from '@/redux/cart/cartSlice';
 import { Empty } from '@/components/Empty';
 import { CartForm } from './CartForm';
 import { CartList } from './CartList';
-import { useAppDispatch, useAppSelector } from '@/redux/hooks';
-import { deleteItem, getFilledCart } from '@/redux/cart/cartSlice';
-import { getProductsAll } from '@/redux/products/productsSlice';
 
 interface CartContentProps {
   deleteAllProducts: () => void;
@@ -14,12 +13,11 @@ export function CartContent({
   deleteAllProducts,
   openModal,
 }: CartContentProps) {
-  const filledCart = useAppSelector(getFilledCart);
-  const productsAll = useAppSelector(getProductsAll)
+  const filteredCart = useAppSelector(getFilteredCart);
 
   const dispatch = useAppDispatch();
 
-  const order: TOrdered = filledCart.map(item => {
+  const order: TOrdered = filteredCart.map(item => {
     return {
       title: item.title,
       quantity: item.quantity,
@@ -30,27 +28,17 @@ export function CartContent({
     dispatch(deleteItem(id));
   };
 
-  const filteredCart =
-    productsAll.filter(({ _id: id1 }) => filledCart.some(({ _id: id2 }) => id1 === id2)
-    );
-  console.log(filteredCart);
-
-  console.log(process.env.NODE_ENV);
+  if (filteredCart.length === 0) {
+    return <Empty text={'Кошик порожній!'} />;
+  }
 
   return (
     <>
-      {filledCart.length > 0 ? (
-        <>
-          <CartList
-            filledCart={filledCart}
-            deleteCartItem={deleteCartItem}
-            deleteAllProducts={deleteAllProducts}
-          />
-          <CartForm openModal={openModal} order={order} />
-        </>
-      ) : (
-        <Empty text={'Кошик порожній!'} />
-      )}
+      <CartList
+        deleteCartItem={deleteCartItem}
+        deleteAllProducts={deleteAllProducts}
+      />
+      <CartForm openModal={openModal} order={order} />
     </>
   );
 }
